Update 2018-11-11-Getting_Started_with_StatsBomb_Data.md
Altered file to _file to avoid clobbering the builtin function. Created a function to generate the list of files allowing that to be used to loop over the subfolders from which lists of files need to be created. As I've indicated, I think the .strip builtin string method is  a clearer way of removing the .json file extension. It's late, so I'm not certain, and don't have time to play around with the data, but I think there's probably a way of reducing the number of times that this section:

data = json.load(data_file)
        #get the nested structure into a dataframe 
        df_ = json_normalize(data, sep = "_")

appears by encapsulating in a function, but it depends on what the downstream stuff is doing exactly. Looks really nice though - and great that it's open source :)